revise model, add some DAO

+++ LabAttendenceSystem/app/src/main/java/sg/edu/ntu/scse/labattendancesystem/domain/models/Main.kt
@@ -2,17 +2,17 @@ package sg.edu.ntu.scse.labattendancesystem.domain.models
 
 import java.time.LocalDateTime
 
-data class Course(val id: Long, val code: String, val title: String)
+data class Course(val id: Int, val code: String, val title: String)
 
 data class GroupStudent(
-    val id: Long,
+    val id: Int,
     val group: Group,
     val student: User,
     val seat: String,
 )
 
 data class Group(
-    val id: Long,
+    val id: Int,
     val name: String,
     val course: Course,
     val lab: Lab,
@@ -23,14 +23,14 @@ data class Group(
 )
 
 data class Session(
-    val id: Long,
+    val id: Int,
     val group: Group,
     val startTime: LocalDateTime,
     val endTime: LocalDateTime
 )
 
 data class Attendance(
-    val id: Long,
+    val id: Int,
     val session: Session,
     val attender: User,
     val checkInState: String,
@@ -39,8 +39,8 @@ data class Attendance(
 )
 
 data class NewAttendance(
-    val sessionId: Long,
-    val attenderId: Long,
+    val sessionId: Int,
+    val attenderId: Int,
     val checkInState: String,
     val checkInDatetime: LocalDateTime,
     val lastModify: LocalDateTime,
