+++ LabAttendenceSystem/app/src/main/java/sg/edu/ntu/scse/labattendancesystem/domain/models/User.kt
@@ -2,13 +2,13 @@ package sg.edu.ntu.scse.labattendancesystem.domain.models
 
 
 data class User(
-    val id: Long,
+    val id: Int,
     val username: String,
     val displayName: String,
 )
 
 data class Lab(
-    val id: Long,
+    val id: Int,
     val username: String,
     val displayName: String,
     val roomCount: Int,
