+++ LabAttendenceSystem/app/src/main/java/sg/edu/ntu/scse/labattendancesystem/network/models/UserResp.kt
@@ -3,7 +3,7 @@ package sg.edu.ntu.scse.labattendancesystem.network.models
 import com.squareup.moshi.Json
 
 data class UserResp(
-    val id: Long? = 0,
+    val id: Int? = 0,
     val username: String? = "",
     @Json(name = "display_name") val displayName: String? = "",
 )
